requirements.txt
# Core requirements - simplified for faster installation
chromadb==0.5.5
sentence-transformers==3.0.1
groq
pyyaml
pydantic>=2.6.0,<3
tqdm
tabulate
pytest

# Minimal agentic framework integration (just what we need)
langchain-core==0.2.43
langchain-text-splitters==0.2.4
